Update bonus header

+++ includes/fdf_bonus.h
@@ -6,7 +6,7 @@
 /*   By: arsobrei <arsobrei@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/25 09:15:50 by arsobrei          #+#    #+#             */
-/*   Updated: 2023/10/17 15:03:09 by arsobrei         ###   ########.fr       */
+/*   Updated: 2023/10/17 15:10:13 by arsobrei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -67,7 +67,7 @@ typedef struct s_map
 	int		width;
 	int		height;
 	float	max_z;
-	int		has_color;
+	short	has_color;
 }			t_map;
 
 typedef struct s_cam
